requirements.txt
django
django-cors-headers
djangorestframework
django-environ
requests